feat: Add kubectl plugin
This starts the dev logs for where we can ingest kubectl events.
From a glance, just looking at the bash history, we don't see extra
context, like cluster and namespace, if those were set in the context
and not specified in the command.

This allows us to hook in there, and track information there for
debugging.