tool: Add scripts for testing local models
We picked qwen3b as a default choice, but without data.

This script runs a matrix of prompts, for different events, on different
models, so that we can figure out what is the best actual final prompt
to use at any given time.

The underlying tool here is Ollama, these are all locally hosted, so
there are no API costs associated with this.