feat: Improve search feature
Previously it was much more minimal in nature but I've since realized a
huge part of this is the ability to go back in time to ANY command, and
get context from that command.

Did it work. What branch were we on. What cluster were we connecting to.

Etc etc, things that are way more useful than just ctrl-r.