fix: Make summaries better by adding exclude list
I find the idea of a clipboard manager that I own and can query helpful.
It was absolutely wrecking the summaries though unfortunately.

We can now exclude plugins / other forms of input as necessary.
We also now have a summarizer backfill command as a dev mode tool, that
can be used to rebuild summaries. My summaries for the day kind of
broke, so I wiped them and rebuilt them.