chore: Backfill method now takes 1 argument
Previously, you needed to have a start and end time.

Theoretically, you can just backfill a small chunk of time.

In reality, it never actually gets used like that, we just ran it for
the day after making some changes to verify them.

So now it just takes 1 arg, which is the day, and defaults to today.