feat: Add repo and branch to more plugins
We're starting to consolidate more around the idea of this ONLY being a
dev journal.

Every developer activity realistically happpens in a repo, with a branch
name.

If I get into a better habit of just switching branches asap, rather
than after the fact when my commit is done, that's a minor amount of
effort needed for so much more context that can be easily used and
hooked into.