feat: Improve logs
We've got 2 kinds of logs, an actual log file that is just taking
everything. Which is great, dig it, maybe in the future I'll want to
update this to not be EVERYTHING but for now I think this is fine.

And our foreground logs are much nicer and easier to read.